chore(cmd): Remove UintVar and move to Uint64Var

+++ internal/cmd/commands/clientagentcmd/clientagentcmd.go
@@ -67,6 +67,6 @@ func (c *ClientAgentCommand) Run(args []string) int {
 }
 
 // clientAgentUrl constructs the full URL for a client agent request given a port and path.
-func clientAgentUrl(port uint, path string) string {
+func clientAgentUrl(port uint64, path string) string {
 	return fmt.Sprintf("http://localhost:%d/%s", port, path)
 }
